refactor: cambia las condicionales de routerAuth.js, validaciones de la comanda
